Fix netlify build script
